fix(web): split client ip API helpers
