+++ web/src/api/index.ts
@@ -1,1 +1,2 @@
 export * from './runtime'
+export * from './clientIp'
